requirements.txt
langchain-openai===0.3.19
langchain-ibm===0.3.19
langchain===0.3.25
openai==1.68.2
python-dotenv==1.0.1